Added NOT NULL constraint to create/update date of Account
The validation annotation, however, was not added in the entity class,
in order to avoid validation when users submit the form. Ideally, this
will be auto-generated and be set from the app side.

The column definition has been marked as non-nullable, with the
createdDate variable being marked as non-updatable as well.