Initialized Common API with Error Response DTOs
In order to properly handle validation error in the application, a
common API has been set up, which - by design - is to take in a Spring
Errors object and convert it into a response DTO that can be transfered
in JSON form.

In this commit, the Spring Context & Lombok dependencies are set up, and
the ErrorResponse DTO is created with basic properties. Along with the
base response DTO, an ErrorDto class is created, which holds the actual
error information in a list, in the response object.