Added Validation Response Handling for Account Type
A ControllerAdvice has been implemented, in which handler method for
MethodArgumentNotValidException has been added, which simply takes the
BindingResult object, passes it to the Common API's Response Utils, and
retrieves the ErrorResponse DTO to be returned to client.

On the AccountTypeController, Valid and RequestBody annotation had to be
added in order to allow for user input to be validated and the exception
to be triggered.