Fixed hibernate mapping for byte[] with PostgreSQL BYTEA
The BYTEA datatype in PostgreSQL expects a "true" byte array, while
hibernate by default seemingly passes an integer representation of the
bytes. This causes PostgreSQL to throw error on persist.

The issue is remedied by using @JdbcTypeCode annotation and Types.BINARY
as the param, to ensure proper casting of the array.