Implemented Supporting Document Retrieval methods
Previously, only document upload functionality was implemented. With
this change, the API now can be accessed to retrieve the uploaded file,
as well as retrieving a list of files or information about a file.

The following end-points have been added:
* /all - GET - Lists all the document uploaded
* /info/{id} - GET - Retrieves information about a file with id (does
  not contain the data)
* /download/{id} - GET - Retrieves the file for download

The upload end-point remains the same:
* /upload - POST - takes in "file" from request and stores it in DB

In addition, some of the issues with the previous code has been
resolved.