Added Content Type information to Supporting Document entity
It is vital to store Content Type information of the file uploaded,
since that is required for proper downloading of the file.

Thus, the column has been added in the "db-scripts.sql" file, along
with the entity.

In addition, the Content Type is now assigned properly while uploading
the file.