Initial Commit for Service Registry
A new project has been set up for implementing the Service Registry of
the project, powered by the Eureka server.

The Spring Boot Web & Spring Boot Eureka Server dependencies have been
added to begin the implementation.

The Spring Dependency Injection plugin is required for this project,
since Spring Cloud Dependencies is required to be injected in order for
Eureka Server to work properly.