Set up Eureka Server for Service Discovery
The Eureka server has been set up in port 8050, with the appropriate
flags put application.properties to mark that it is an Eureka server,
not a client (by default, Eureka considers itself as a client).

In addition, the AppRunner class has been set up as the entry point of
the Service Registry, with the @EnableEurekaServer annotation on top to
indicate (and interally configure itself accordingly) it is an Eureka Server.