Configured Account API to be registered to Eurkea Server
The process of configuring the API was simply to add the following
dependencies:

* Spring Dependency Injection plugin
* Spring Eureka Client Starter
* Spring Cloud Dependenices (as a mavenBom)

In addition, the API service was given a name - account-api, in this
case - and the service url for Eureka Server was defined.

Note: As advised by many articles, @EnableEurekaClient annotation is
required to be put on top of the main class. However, this is no longer
required, since the annotation has since been deprecated.