Added Configuration to APIs to show api name in Eureka
By default, in the Eureka dashboard, the machine id along with port is
shown under the Status column, as a link to the actuator of the service.
Also, the link includes the machine id, instead of localhost or local
IP.

This isn't ideal, since we may not want to expose the name in the
dashboard. As such, configuration has been shown to show API name under
the Status column and the link to be constructed with the machine IP.