Added SupportingDocumentService Feign Client in Account API
In order to access the Supporting Document API from the Account API, a
FeignClient is created, which refers to the required endpoints of the
API.

The functionalities include uploading and downloading files, removing
them, along with retrieving information of the file.

The FeignClient annotation contains the service name, along with the
context path, in order to make the requests properly.