Added document_id column to account table
The column name has been added to the table creation script under
resources/db-scripts.sql.

The entity has also been modified to house the variable for the column,
along with a transient SupportingDocumentDto variable.

The idea is to store just the id of the document in the table, which
will be then used to retrieve the SupportingDocumentDto whenever required.