Added required dependencies for Gateway Server
Spring Netflix Zuul dependency has been added to achieve the purpose of
implementing the gateway server. Along with the Zuul dependency, Spring
Boot Web Starter & Spring Eureka Client dependency has been added, along
with the Spring Cloud dependencies.

It should be noted that the @EnableZuulProxy annotation did not work
with the latest version of Spring Boot. The version used was 2.7.7,
which is seemingly the last version where Zuul works perfectly.

In addition, to supporting the older version of Spring Boot and Zuul,
Spring Cloud Dependencies 2021.0.6 is being used.