Configured Auth API to be registered to Eureka Server
In order to make the Auth API eligible for register with Eureka Server,
the Eureka client dependency was added, along with the cloud plugin.

The "auth-api" name has been assigned as application name and Eureka
instance id, which will be used for API discovery in Eureka and register
the API in the gateway server.