Added latest GSON dependency to Gateway build script
When starting the Gateway Server, the service would throw an exception,
resulting the start up to fail. The exception stated that it could not
set a flag in the GSON library.

Upon further digging, it seems that the version of GSON included in the
Spring Cloud Gateway do not contain the flag intended to be updated.
Thus, the module had to be excluded from the Spring Cloud Gateway
dependency, in favor of adding the latest version manually to
build.gradle.