Fixed issues with Login & Login Response Generation
As per the application design, the roles set in Login entity must have
Enumerated annotation placed on top, indicating that the Role enum's
String would be stored in the database - not the ordinal.

Also, there was an ordering mismatch in ResponseUtils when generating
the LoginResponseDto. The status information was put in place of Last
Name, causing the DTO to be populated incorrectly.

Both of these have been fixed.