Implemented LoginDetailService in User API
Spring Security uses it's own User object to authenticate user
internally. As such, there is a need to convert the Login entity object
to User object.

The LoginDetailService is created to resolve the issue - it uses the
LoginService to retrieve the Login using username (one provided by the
user in the request) and then converts it to Spring's User object.

If the login is not found, a NoResultException is thrown.