Added AuthenticationProvider bean to SecurityConfig
In order to ensure that Spring Security uses the LoginDetailService and
the PasswordEncoder created for the API, the AuthenticationProvider bean
is created.

The bean is represented by a DaoAuthenticationProvider instance, where
the LoginDetailService & PasswordEncoder beans are injected - the latter
being defined in the AppRunner class, since it is used across multiple
classes.