Added JwtUsernamePasswordAuthFilter to User API
The filter that will authenticate user based on the credentials entered
and return a token has been implemented. JwtUsernamePasswordAuthFilter
takes in username and password from the request and tries to
authenticate the user in the attemptAuthentication method.

If the attempt was successful, the token is generated inside the
successfulAuthentication method. The token is then returned in the
response.

If the attempt was unsuccessful, a 401 error is thrown as the response.

The filter is added in the config class, so that it can execute on every
incoming request to the registered path.

Also, the API key is kept in the class. This will later be moved to a
properties file to maintain integrity and security of the key.