Fixed issue with lazy loading roles in Login
When authenticating user, the roles list of Login is needed to be
accessed.

Since the roles variable was marked to be lazily loaded, the
authentication process failed since it only found the proxy of the list.

The list had to be eagerly fetched in order for the authentication to
work properly.