Added /user/error as exposed URL in SecurityConfig
In order to display error after failed authentication, the /user/error URL
had to be added to the whitelist, permitting all requests to pass to it.

Out of the box, Spring uses BasicErrorController to handle errors. The
default endpoint for this controller "/error", which violates the URL
pattern of the api. Thus, the "/user/error" path has defined to
application.properties file, overriding the default behavior.