Moved User Authentication Secret to resources file
Till now, for testing purposes, a temporary key was generated and was
placed in SecurityConfig class, which was used to generate the token on
authentication.

This is a bad practice, since secret keys should not be at plain sight.
Thus, a new file has been introduced in the resources folder, namely
"secret.properties". The file contains the secret that can be referenced
with the key "api.auth.secret".

Using @PropertyResource annotation to identify the file and @Value
annotation to assign the value to a variable in SecurityConfig, the key
has been put to use from the resource file.