Added AuthController to User API
In order to isolate the endpoints that are accessible by all, a new
controller has been introduced, namely AuthController. Currently, there
is only one endpoint - "/save" - which takes in a LoginRequestDto and
saves the login information entered by the user.

The "/save" method was originally in LoginController and was explictly
kept exposed in the SecurityConfig. However, this added an extra bit of
complexity, since the "/authenticate" endpoint has a prefix of "/auth",
while the "/save" endpoint has the endpoint of "/login".

In order to generalize the path prefix and keep it exposed to all at all
times, "/auth" prefix has been chosen, which resulted in the creation of
the controller. In time, the GET mapping to navigate to the view pages
of the login & register will be added here.

Also, with the change, a typo has been fixed in the JwtUsernamePasswordFilter.