Setup initial Security Config in Gateway Server
SecurityConfig class has been introduced in the Gateway Server, which
ensures that all requests for "/user/auth/**" passed without authentication.
However, all other requests need to be authenticated in order to access
the resources.

The session creation policy has been set as stateless, while 401 error
is to be thrown when an error occurs at the authentication endpoint.

Note that the requests will be authenticated by roles in the future. For
now, they simply need to be authenticated to access the resource, except
for the auth endpoints.