Fixed issue with extracting claims in Common API
In order to properly parse the token, the token parser requires the
secret key that was used to generate it.

As such, the getClaims method in JwtUtils now has a new parameter -
secret - which is to be sent from the API that is seeking to parse
the token.

Also, removed unused empty constant API_SECRET from the said utils
class.