Added secret key to SecurityConfig of Gateway Server
In order to be able to parse the token in the request, the key which was
used to generate the token is needed. Thus, the secret.properties file
has been added to the Gateway Server's resources directory.

However, since the key is sensitive in nature, tweaks have been made to
the gitignore file to ensure that it is not included in the git's list
of tracked files.

The SecurityConfig class has been modified as well to hold the key
variable, which is resovled via PropertyResources annotation placed at
the top of the class.