Refactored Login Request DTOs
Initially, LoginRequestDto class was introduced in the Common API,
allowing for using the same class to create objects to hold data for
both login and registration.

However, this allows user more flexibility, as users could potentially
be able to alter their data while logging in. This isn't ideal, and thus,
the initial simplicity of the approach turned out to be a future
potential bug.

Thus, the DTOs for login and registration has been separated to two
separate classes. In the Common API, LoginRequestDto is retained, which
now only takes in user inputs for username and password.

A new LoginRegistrationDto has been introduced in the User API, which
holds all the properties needed for registration. This was not put in
the Common API, since the registration process will solely be handlded
by the User API.